# openapi.yaml
openapi: 3.0.3
info:
  title: Digital Insurance API
  version: 1.0.0
  description: >-
    Customer-facing API for the Digital Insurance platform. Handles quotes,
    claims, and customers; binds and prices policies through the internal Policy
    API; and charges premiums and fetches payment receipts through the internal
    Payment API.
servers:
  - url: https://api.digitalinsurance.dev
    description: Production
paths:
...
  /v1/policies:
    post:
      operationId: bindPolicy
      summary: Bind a policy (creates and underwrites it via the Policy API)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BindPolicyRequest'
      responses:
        '201':
          description: >-
            Policy bound. Created via POST /v1/policies and underwritten via
            POST /v1/policies/{policyId}/underwrite on the internal Policy API.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Policy'
  /v1/policies/{policyId}:
    get:
      operationId: getPolicy
      summary: Fetch a bound policy from the Policy API
      parameters:
        - name: policyId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: >-
            Policy found. Retrieved via GET /v1/policies/{policyId} on the
            internal Policy API.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Policy'
        '404':
          description: Policy not found
  /v1/policies/{policyId}/pay:
    post:
      operationId: payPolicy
      summary: Pay the policy premium via the Payment API
      parameters:
        - name: policyId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PayRequest'
      responses:
        '200':
          description: >-
            Premium charged via POST /v1/payments/charge on the internal Payment
            API.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Payment'
...
components:
  schemas:
...
    BindPolicyRequest:
      type: object
      required:
        - quoteId
        - customerId
        - paymentMethodId
      properties:
        quoteId:
          type: string
        customerId:
          type: string
        paymentMethodId:
          type: string
    Policy:
      type: object
      properties:
        id:
          type: string
        quoteId:
          type: string
        customerId:
          type: string
        status:
          type: string
        premiumCents:
          type: integer
    PayRequest:
      type: object
      required:
        - paymentMethodId
      properties:
        paymentMethodId:
          type: string
...
    Payment:
      type: object
      properties:
        id:
          type: string
        policyId:
          type: string
        status:
          type: string
        amountCents:
          type: integer
        currency:
          type: string
        createdAt:
          type: string
...